requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

Pillow == 9.2.0
PyYAML == 6.0
einops == 0.4.1
gradio == 3.2
hydra_core == 1.2.0
hydra-colorlog == 1.2.0
hydra-optuna-sweeper == 1.2.0
jieba == 0.42.1
kiwisolver == 1.4.4
matplotlib == 3.5.3
numpy == 1.23.2
omegaconf == 2.2.3
packaging == 21.3
pandas == 1.4.4
pyrootutils == 1.0.4
pytest == 7.1.2
pytorch_lightning == 1.7.0
requests == 2.28.1
rich == 12.5.1
scikit_learn == 1.1.2
setuptools == 65.3.0
sh == 1.14.3
torch == 1.12.1
torchmetrics == 0.9.2
torchvision == 0.13.1
torchtext == 0.13.1
tqdm == 4.64.0
transformers == 4.21.2
wandb == 0.13.2
pre-commit
